+++ src/common/header/Header.jsx
@@ -1,14 +1,14 @@
 import React from 'react'
-import * as S from './HerderStyle'
+import * as S from './HeaderStyle'
 import { AiFillShopping } from 'react-icons/ai'
 import { useNavigate } from 'react-router-dom'
 
-function Herder() {
+function Header() {
 
   const navigate = useNavigate()
 
   return (
-    <S.Herder>
+    <S.Header>
       <S.Container>
         <S.Flex>
           <S.Img src="https://cdn.sisamagazine.co.kr/news/photo/202109/401832_401982_133.png" alt="" />
@@ -18,8 +18,8 @@ function Herder() {
           </S.LoginTextContainer>
         </S.Flex>
       </S.Container>
-    </S.Herder>
+    </S.Header>
   )
 }
 
-export default Herder
+export default Header
