+++ src/common/header/HeaderStyle.js
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 
-export const Herder = styled.div `
+export const Header = styled.div `
   width: 425px;
   height: 8vh;
   margin-bottom: 2em;
